--- res/sources/first_source.yaml
@@ -1,5 +1,0 @@
-- flows:
-  - data_rate: 1.0
-    id: f1
-  node: pop0
-  vnf: vnf_source
