--- res/sources/source.yaml
@@ -1,1 +1,0 @@
-[]
