+++ results/abilene-in1-cap1/abc/det-arrival10_det-size001_duration06/2019-11-28_16-03-11_seed6043/abc.yaml
@@ -1,0 +1,20 @@
+# simple chain of 3 SFs a->b->c with deterministic processing delays
+
+# list of SFCs and involved SFs (order of SFs matters). names need to match dummy schedule and placement (dummy_data.py)
+sfc_list:
+  sfc_1:
+    - a
+    - b
+    - c
+
+# SF attributes (for now, processing delay)
+sf_list:
+  a:
+    processing_delay_mean: 5.0
+    processing_delay_stdev: 0.0
+  b:
+    processing_delay_mean: 5.0
+    processing_delay_stdev: 0.0
+  c:
+    processing_delay_mean: 5.0
+    processing_delay_stdev: 0.0
