+++ results/abilene-in1-cap1/abc/det-arrival10_det-size001_duration06/2019-11-28_16-03-11_seed6043/det-arrival10_det-size001_duration06.yaml
@@ -1,0 +1,11 @@
+# module for configuring the simulator
+# configuration parameters are loaded and used both when using the simulator via the CLI and via the interface
+# all parameters are required, defaults are in comments
+
+inter_arrival_mean: 10.0           # default: 10.0
+deterministic_arrival: True
+flow_dr_mean: 1.0                  # default: 1.0
+flow_dr_stdev: 0.0                 # default: 0.0
+flow_size_shape: 0.001             # default: 0.001
+deterministic_size: True
+run_duration: 6                  # default: 100
