--- results/bjointsp/triangle-sfc_1-first_source-2019-08-02_18-10-07_2034.yaml
@@ -1,127 +1,0 @@
-input:
-  algorithm: bjointsp
-  fixed: None
-  network: triangle.graphml
-  num_edges: 3
-  num_nodes: 3
-  num_sources: 1
-  num_vnfs: 4
-  objective: 0
-  prev_embedding: None
-  seed: 2034
-  service: sfc_1.yaml
-  sources: first_source.yaml
-metrics:
-  changed:
-  - name: c
-    node: pop0
-  - name: a
-    node: pop0
-  - name: b
-    node: pop0
-  - name: vnf_source
-    node: pop0
-  max_cpu_oversub: 0
-  max_dr_oversub: 0
-  max_endToEnd_delay: 15.0
-  max_mem_oversub: 0
-  num_changed: 4
-  num_instances: 4
-  obj_value: 4003015.0
-  path_delays:
-  - dest: c
-    dest_node: pop0
-    path_delay: 0
-    src: b
-    src_node: pop0
-  - dest: b
-    dest_node: pop0
-    path_delay: 0
-    src: a
-    src_node: pop0
-  - dest: a
-    dest_node: pop0
-    path_delay: 0
-    src: vnf_source
-    src_node: pop0
-  runtime: 0.052834272384643555
-  total_delay: 15.0
-  total_path_delay: 0
-  total_vnf_delay: 15.0
-  vnf_delays:
-  - vnf: c
-    vnf_delay: 5.0
-  - vnf: a
-    vnf_delay: 5.0
-  - vnf: b
-    vnf_delay: 5.0
-  - vnf: vnf_source
-    vnf_delay: 0
-placement:
-  alloc_node_res:
-  - cpu: 1.0
-    mem: 0.0
-    name: c
-    node: pop0
-  - cpu: 1.0
-    mem: 0.0
-    name: a
-    node: pop0
-  - cpu: 1.0
-    mem: 0.0
-    name: b
-    node: pop0
-  - cpu: 0
-    mem: 0
-    name: vnf_source
-    node: pop0
-  cpu_oversub: []
-  dr_oversub: []
-  flows:
-  - arc: b.0->c.0
-    dest_vnf: c
-    dst_node: pop0
-    flow_id: f1
-    src_node: pop0
-    src_vnf: b
-  - arc: a.0->b.0
-    dest_vnf: b
-    dst_node: pop0
-    flow_id: f1
-    src_node: pop0
-    src_vnf: a
-  - arc: vnf_source.0->a.0
-    dest_vnf: a
-    dst_node: pop0
-    flow_id: f1
-    src_node: pop0
-    src_vnf: vnf_source
-  links: []
-  mem_oversub: []
-  vlinks:
-  - dest_node: pop0
-    dest_vnf: c
-    src_node: pop0
-    src_vnf: b
-  - dest_node: pop0
-    dest_vnf: b
-    src_node: pop0
-    src_vnf: a
-  - dest_node: pop0
-    dest_vnf: a
-    src_node: pop0
-    src_vnf: vnf_source
-  vnfs:
-  - image: null
-    name: c
-    node: pop0
-  - image: null
-    name: a
-    node: pop0
-  - image: null
-    name: b
-    node: pop0
-  - image: null
-    name: vnf_source
-    node: pop0
-time: 2019-08-02_18-10-07
