--- results/metrics-2019-08-02_18-10-07.csv
@@ -1,11 +1,0 @@
-time,total_flows,successful_flows,dropped_flows,in_network_flows,avg_end_2_end_delay
-100.0,10,9,0,1,15.0
-200.0,20,19,0,1,15.0
-300.0,30,29,0,1,15.0
-400.0,40,39,0,1,15.0
-500.0,50,49,0,1,15.0
-600.0,60,59,0,1,15.0
-700.0,70,69,0,1,15.0
-800.0,80,79,0,1,15.0
-900.0,90,89,0,1,15.0
-1000.0,100,99,0,1,15.0
