--- results/placements-2019-08-02_18-10-07.csv
@@ -1,41 +1,0 @@
-time,node,sf
-0,pop0,c
-0,pop0,a
-0,pop0,b
-0,pop0,vnf_source
-100.0,pop0,c
-100.0,pop0,a
-100.0,pop0,b
-100.0,pop0,vnf_source
-200.0,pop0,c
-200.0,pop0,a
-200.0,pop0,b
-200.0,pop0,vnf_source
-300.0,pop0,c
-300.0,pop0,a
-300.0,pop0,b
-300.0,pop0,vnf_source
-400.0,pop0,c
-400.0,pop0,a
-400.0,pop0,b
-400.0,pop0,vnf_source
-500.0,pop0,c
-500.0,pop0,a
-500.0,pop0,b
-500.0,pop0,vnf_source
-600.0,pop0,c
-600.0,pop0,a
-600.0,pop0,b
-600.0,pop0,vnf_source
-700.0,pop0,c
-700.0,pop0,a
-700.0,pop0,b
-700.0,pop0,vnf_source
-800.0,pop0,c
-800.0,pop0,a
-800.0,pop0,b
-800.0,pop0,vnf_source
-900.0,pop0,c
-900.0,pop0,a
-900.0,pop0,b
-900.0,pop0,vnf_source
