--- results/resources-2019-08-02_18-10-07.csv
@@ -1,31 +1,0 @@
-time,node,node_capacity,used_resources
-100.0,pop0,10,1.0
-100.0,pop1,10,0
-100.0,pop2,10,0
-200.0,pop0,10,1.0
-200.0,pop1,10,0
-200.0,pop2,10,0
-300.0,pop0,10,1.0
-300.0,pop1,10,0
-300.0,pop2,10,0
-400.0,pop0,10,1.0
-400.0,pop1,10,0
-400.0,pop2,10,0
-500.0,pop0,10,1.0
-500.0,pop1,10,0
-500.0,pop2,10,0
-600.0,pop0,10,1.0
-600.0,pop1,10,0
-600.0,pop2,10,0
-700.0,pop0,10,1.0
-700.0,pop1,10,0
-700.0,pop2,10,0
-800.0,pop0,10,1.0
-800.0,pop1,10,0
-800.0,pop2,10,0
-900.0,pop0,10,1.0
-900.0,pop1,10,0
-900.0,pop2,10,0
-1000.0,pop0,10,1.0
-1000.0,pop1,10,0
-1000.0,pop2,10,0
