--- results/scheduling_2019-08-02_18-10-07.csv
@@ -1,271 +1,0 @@
-time,origin_node,sfc,sf,schedule_node,schedule_prob
-0,pop0,sfc_1,a,pop0,1
-0,pop0,sfc_1,a,pop1,0
-0,pop0,sfc_1,a,pop2,0
-0,pop0,sfc_1,b,pop0,1
-0,pop0,sfc_1,b,pop1,0
-0,pop0,sfc_1,b,pop2,0
-0,pop0,sfc_1,c,pop0,1
-0,pop0,sfc_1,c,pop1,0
-0,pop0,sfc_1,c,pop2,0
-0,pop1,sfc_1,a,pop0,0.34
-0,pop1,sfc_1,a,pop1,0.33
-0,pop1,sfc_1,a,pop2,0.33
-0,pop1,sfc_1,b,pop0,0.34
-0,pop1,sfc_1,b,pop1,0.33
-0,pop1,sfc_1,b,pop2,0.33
-0,pop1,sfc_1,c,pop0,0.34
-0,pop1,sfc_1,c,pop1,0.33
-0,pop1,sfc_1,c,pop2,0.33
-0,pop2,sfc_1,a,pop0,0.34
-0,pop2,sfc_1,a,pop1,0.33
-0,pop2,sfc_1,a,pop2,0.33
-0,pop2,sfc_1,b,pop0,0.34
-0,pop2,sfc_1,b,pop1,0.33
-0,pop2,sfc_1,b,pop2,0.33
-0,pop2,sfc_1,c,pop0,0.34
-0,pop2,sfc_1,c,pop1,0.33
-0,pop2,sfc_1,c,pop2,0.33
-100.0,pop0,sfc_1,a,pop0,1
-100.0,pop0,sfc_1,a,pop1,0
-100.0,pop0,sfc_1,a,pop2,0
-100.0,pop0,sfc_1,b,pop0,1
-100.0,pop0,sfc_1,b,pop1,0
-100.0,pop0,sfc_1,b,pop2,0
-100.0,pop0,sfc_1,c,pop0,1
-100.0,pop0,sfc_1,c,pop1,0
-100.0,pop0,sfc_1,c,pop2,0
-100.0,pop1,sfc_1,a,pop0,0.34
-100.0,pop1,sfc_1,a,pop1,0.33
-100.0,pop1,sfc_1,a,pop2,0.33
-100.0,pop1,sfc_1,b,pop0,0.34
-100.0,pop1,sfc_1,b,pop1,0.33
-100.0,pop1,sfc_1,b,pop2,0.33
-100.0,pop1,sfc_1,c,pop0,0.34
-100.0,pop1,sfc_1,c,pop1,0.33
-100.0,pop1,sfc_1,c,pop2,0.33
-100.0,pop2,sfc_1,a,pop0,0.34
-100.0,pop2,sfc_1,a,pop1,0.33
-100.0,pop2,sfc_1,a,pop2,0.33
-100.0,pop2,sfc_1,b,pop0,0.34
-100.0,pop2,sfc_1,b,pop1,0.33
-100.0,pop2,sfc_1,b,pop2,0.33
-100.0,pop2,sfc_1,c,pop0,0.34
-100.0,pop2,sfc_1,c,pop1,0.33
-100.0,pop2,sfc_1,c,pop2,0.33
-200.0,pop0,sfc_1,a,pop0,1
-200.0,pop0,sfc_1,a,pop1,0
-200.0,pop0,sfc_1,a,pop2,0
-200.0,pop0,sfc_1,b,pop0,1
-200.0,pop0,sfc_1,b,pop1,0
-200.0,pop0,sfc_1,b,pop2,0
-200.0,pop0,sfc_1,c,pop0,1
-200.0,pop0,sfc_1,c,pop1,0
-200.0,pop0,sfc_1,c,pop2,0
-200.0,pop1,sfc_1,a,pop0,0.34
-200.0,pop1,sfc_1,a,pop1,0.33
-200.0,pop1,sfc_1,a,pop2,0.33
-200.0,pop1,sfc_1,b,pop0,0.34
-200.0,pop1,sfc_1,b,pop1,0.33
-200.0,pop1,sfc_1,b,pop2,0.33
-200.0,pop1,sfc_1,c,pop0,0.34
-200.0,pop1,sfc_1,c,pop1,0.33
-200.0,pop1,sfc_1,c,pop2,0.33
-200.0,pop2,sfc_1,a,pop0,0.34
-200.0,pop2,sfc_1,a,pop1,0.33
-200.0,pop2,sfc_1,a,pop2,0.33
-200.0,pop2,sfc_1,b,pop0,0.34
-200.0,pop2,sfc_1,b,pop1,0.33
-200.0,pop2,sfc_1,b,pop2,0.33
-200.0,pop2,sfc_1,c,pop0,0.34
-200.0,pop2,sfc_1,c,pop1,0.33
-200.0,pop2,sfc_1,c,pop2,0.33
-300.0,pop0,sfc_1,a,pop0,1
-300.0,pop0,sfc_1,a,pop1,0
-300.0,pop0,sfc_1,a,pop2,0
-300.0,pop0,sfc_1,b,pop0,1
-300.0,pop0,sfc_1,b,pop1,0
-300.0,pop0,sfc_1,b,pop2,0
-300.0,pop0,sfc_1,c,pop0,1
-300.0,pop0,sfc_1,c,pop1,0
-300.0,pop0,sfc_1,c,pop2,0
-300.0,pop1,sfc_1,a,pop0,0.34
-300.0,pop1,sfc_1,a,pop1,0.33
-300.0,pop1,sfc_1,a,pop2,0.33
-300.0,pop1,sfc_1,b,pop0,0.34
-300.0,pop1,sfc_1,b,pop1,0.33
-300.0,pop1,sfc_1,b,pop2,0.33
-300.0,pop1,sfc_1,c,pop0,0.34
-300.0,pop1,sfc_1,c,pop1,0.33
-300.0,pop1,sfc_1,c,pop2,0.33
-300.0,pop2,sfc_1,a,pop0,0.34
-300.0,pop2,sfc_1,a,pop1,0.33
-300.0,pop2,sfc_1,a,pop2,0.33
-300.0,pop2,sfc_1,b,pop0,0.34
-300.0,pop2,sfc_1,b,pop1,0.33
-300.0,pop2,sfc_1,b,pop2,0.33
-300.0,pop2,sfc_1,c,pop0,0.34
-300.0,pop2,sfc_1,c,pop1,0.33
-300.0,pop2,sfc_1,c,pop2,0.33
-400.0,pop0,sfc_1,a,pop0,1
-400.0,pop0,sfc_1,a,pop1,0
-400.0,pop0,sfc_1,a,pop2,0
-400.0,pop0,sfc_1,b,pop0,1
-400.0,pop0,sfc_1,b,pop1,0
-400.0,pop0,sfc_1,b,pop2,0
-400.0,pop0,sfc_1,c,pop0,1
-400.0,pop0,sfc_1,c,pop1,0
-400.0,pop0,sfc_1,c,pop2,0
-400.0,pop1,sfc_1,a,pop0,0.34
-400.0,pop1,sfc_1,a,pop1,0.33
-400.0,pop1,sfc_1,a,pop2,0.33
-400.0,pop1,sfc_1,b,pop0,0.34
-400.0,pop1,sfc_1,b,pop1,0.33
-400.0,pop1,sfc_1,b,pop2,0.33
-400.0,pop1,sfc_1,c,pop0,0.34
-400.0,pop1,sfc_1,c,pop1,0.33
-400.0,pop1,sfc_1,c,pop2,0.33
-400.0,pop2,sfc_1,a,pop0,0.34
-400.0,pop2,sfc_1,a,pop1,0.33
-400.0,pop2,sfc_1,a,pop2,0.33
-400.0,pop2,sfc_1,b,pop0,0.34
-400.0,pop2,sfc_1,b,pop1,0.33
-400.0,pop2,sfc_1,b,pop2,0.33
-400.0,pop2,sfc_1,c,pop0,0.34
-400.0,pop2,sfc_1,c,pop1,0.33
-400.0,pop2,sfc_1,c,pop2,0.33
-500.0,pop0,sfc_1,a,pop0,1
-500.0,pop0,sfc_1,a,pop1,0
-500.0,pop0,sfc_1,a,pop2,0
-500.0,pop0,sfc_1,b,pop0,1
-500.0,pop0,sfc_1,b,pop1,0
-500.0,pop0,sfc_1,b,pop2,0
-500.0,pop0,sfc_1,c,pop0,1
-500.0,pop0,sfc_1,c,pop1,0
-500.0,pop0,sfc_1,c,pop2,0
-500.0,pop1,sfc_1,a,pop0,0.34
-500.0,pop1,sfc_1,a,pop1,0.33
-500.0,pop1,sfc_1,a,pop2,0.33
-500.0,pop1,sfc_1,b,pop0,0.34
-500.0,pop1,sfc_1,b,pop1,0.33
-500.0,pop1,sfc_1,b,pop2,0.33
-500.0,pop1,sfc_1,c,pop0,0.34
-500.0,pop1,sfc_1,c,pop1,0.33
-500.0,pop1,sfc_1,c,pop2,0.33
-500.0,pop2,sfc_1,a,pop0,0.34
-500.0,pop2,sfc_1,a,pop1,0.33
-500.0,pop2,sfc_1,a,pop2,0.33
-500.0,pop2,sfc_1,b,pop0,0.34
-500.0,pop2,sfc_1,b,pop1,0.33
-500.0,pop2,sfc_1,b,pop2,0.33
-500.0,pop2,sfc_1,c,pop0,0.34
-500.0,pop2,sfc_1,c,pop1,0.33
-500.0,pop2,sfc_1,c,pop2,0.33
-600.0,pop0,sfc_1,a,pop0,1
-600.0,pop0,sfc_1,a,pop1,0
-600.0,pop0,sfc_1,a,pop2,0
-600.0,pop0,sfc_1,b,pop0,1
-600.0,pop0,sfc_1,b,pop1,0
-600.0,pop0,sfc_1,b,pop2,0
-600.0,pop0,sfc_1,c,pop0,1
-600.0,pop0,sfc_1,c,pop1,0
-600.0,pop0,sfc_1,c,pop2,0
-600.0,pop1,sfc_1,a,pop0,0.34
-600.0,pop1,sfc_1,a,pop1,0.33
-600.0,pop1,sfc_1,a,pop2,0.33
-600.0,pop1,sfc_1,b,pop0,0.34
-600.0,pop1,sfc_1,b,pop1,0.33
-600.0,pop1,sfc_1,b,pop2,0.33
-600.0,pop1,sfc_1,c,pop0,0.34
-600.0,pop1,sfc_1,c,pop1,0.33
-600.0,pop1,sfc_1,c,pop2,0.33
-600.0,pop2,sfc_1,a,pop0,0.34
-600.0,pop2,sfc_1,a,pop1,0.33
-600.0,pop2,sfc_1,a,pop2,0.33
-600.0,pop2,sfc_1,b,pop0,0.34
-600.0,pop2,sfc_1,b,pop1,0.33
-600.0,pop2,sfc_1,b,pop2,0.33
-600.0,pop2,sfc_1,c,pop0,0.34
-600.0,pop2,sfc_1,c,pop1,0.33
-600.0,pop2,sfc_1,c,pop2,0.33
-700.0,pop0,sfc_1,a,pop0,1
-700.0,pop0,sfc_1,a,pop1,0
-700.0,pop0,sfc_1,a,pop2,0
-700.0,pop0,sfc_1,b,pop0,1
-700.0,pop0,sfc_1,b,pop1,0
-700.0,pop0,sfc_1,b,pop2,0
-700.0,pop0,sfc_1,c,pop0,1
-700.0,pop0,sfc_1,c,pop1,0
-700.0,pop0,sfc_1,c,pop2,0
-700.0,pop1,sfc_1,a,pop0,0.34
-700.0,pop1,sfc_1,a,pop1,0.33
-700.0,pop1,sfc_1,a,pop2,0.33
-700.0,pop1,sfc_1,b,pop0,0.34
-700.0,pop1,sfc_1,b,pop1,0.33
-700.0,pop1,sfc_1,b,pop2,0.33
-700.0,pop1,sfc_1,c,pop0,0.34
-700.0,pop1,sfc_1,c,pop1,0.33
-700.0,pop1,sfc_1,c,pop2,0.33
-700.0,pop2,sfc_1,a,pop0,0.34
-700.0,pop2,sfc_1,a,pop1,0.33
-700.0,pop2,sfc_1,a,pop2,0.33
-700.0,pop2,sfc_1,b,pop0,0.34
-700.0,pop2,sfc_1,b,pop1,0.33
-700.0,pop2,sfc_1,b,pop2,0.33
-700.0,pop2,sfc_1,c,pop0,0.34
-700.0,pop2,sfc_1,c,pop1,0.33
-700.0,pop2,sfc_1,c,pop2,0.33
-800.0,pop0,sfc_1,a,pop0,1
-800.0,pop0,sfc_1,a,pop1,0
-800.0,pop0,sfc_1,a,pop2,0
-800.0,pop0,sfc_1,b,pop0,1
-800.0,pop0,sfc_1,b,pop1,0
-800.0,pop0,sfc_1,b,pop2,0
-800.0,pop0,sfc_1,c,pop0,1
-800.0,pop0,sfc_1,c,pop1,0
-800.0,pop0,sfc_1,c,pop2,0
-800.0,pop1,sfc_1,a,pop0,0.34
-800.0,pop1,sfc_1,a,pop1,0.33
-800.0,pop1,sfc_1,a,pop2,0.33
-800.0,pop1,sfc_1,b,pop0,0.34
-800.0,pop1,sfc_1,b,pop1,0.33
-800.0,pop1,sfc_1,b,pop2,0.33
-800.0,pop1,sfc_1,c,pop0,0.34
-800.0,pop1,sfc_1,c,pop1,0.33
-800.0,pop1,sfc_1,c,pop2,0.33
-800.0,pop2,sfc_1,a,pop0,0.34
-800.0,pop2,sfc_1,a,pop1,0.33
-800.0,pop2,sfc_1,a,pop2,0.33
-800.0,pop2,sfc_1,b,pop0,0.34
-800.0,pop2,sfc_1,b,pop1,0.33
-800.0,pop2,sfc_1,b,pop2,0.33
-800.0,pop2,sfc_1,c,pop0,0.34
-800.0,pop2,sfc_1,c,pop1,0.33
-800.0,pop2,sfc_1,c,pop2,0.33
-900.0,pop0,sfc_1,a,pop0,1
-900.0,pop0,sfc_1,a,pop1,0
-900.0,pop0,sfc_1,a,pop2,0
-900.0,pop0,sfc_1,b,pop0,1
-900.0,pop0,sfc_1,b,pop1,0
-900.0,pop0,sfc_1,b,pop2,0
-900.0,pop0,sfc_1,c,pop0,1
-900.0,pop0,sfc_1,c,pop1,0
-900.0,pop0,sfc_1,c,pop2,0
-900.0,pop1,sfc_1,a,pop0,0.34
-900.0,pop1,sfc_1,a,pop1,0.33
-900.0,pop1,sfc_1,a,pop2,0.33
-900.0,pop1,sfc_1,b,pop0,0.34
-900.0,pop1,sfc_1,b,pop1,0.33
-900.0,pop1,sfc_1,b,pop2,0.33
-900.0,pop1,sfc_1,c,pop0,0.34
-900.0,pop1,sfc_1,c,pop1,0.33
-900.0,pop1,sfc_1,c,pop2,0.33
-900.0,pop2,sfc_1,a,pop0,0.34
-900.0,pop2,sfc_1,a,pop1,0.33
-900.0,pop2,sfc_1,a,pop2,0.33
-900.0,pop2,sfc_1,b,pop0,0.34
-900.0,pop2,sfc_1,b,pop1,0.33
-900.0,pop2,sfc_1,b,pop2,0.33
-900.0,pop2,sfc_1,c,pop0,0.34
-900.0,pop2,sfc_1,c,pop1,0.33
-900.0,pop2,sfc_1,c,pop2,0.33
