+++ src/adapter/adapter.py
@@ -1,19 +1,20 @@
 import argparse
 import logging
 import os
-
 import yaml
+import random
 from datetime import datetime
 from bjointsp.main import place as bjointsp_place
 from siminterface.simulator import Simulator
 from spinterface.spinterface import SimulatorAction
 
-from yaml_reader.results_reader import get_placement_and_schedule
-from yaml_writer.writer import create_template, create_source_file
+from util.reader import get_placement_and_schedule, get_project_root
+from util.writer import create_template, create_source_file, copy_input_files
 
 log = logging.getLogger(__name__)
 
 BJOINTSP_FIRST_SRC_LOCATION = "res/sources/first_source.yaml"
+DATETIME = datetime.now().strftime("%Y-%m-%d_%H-%M-%S")
 
 
 def get_ingress_nodes_and_cap(network):
@@ -39,7 +40,7 @@ def get_ingress_nodes_and_cap(network):
 def parse_args():
     parser = argparse.ArgumentParser(description="BJointSP Adapter")
     parser.add_argument('-i', '--iterations', required=False, default=10, dest="iterations", type=int)
-    parser.add_argument('-s', '--seed', required=False, default=9999, dest="seed", type=int)
+    parser.add_argument('-s', '--seed', required=False, dest="seed", type=int)
     parser.add_argument('-n', '--network', required=True, dest='network')
     parser.add_argument('-sf', '--service_functions', required=True, dest="service_functions")
     parser.add_argument('-c', '--config', required=True, dest="config")
@@ -48,18 +49,25 @@ def parse_args():
 
 def main():
     args = parse_args()
-    timestamp = datetime.now().strftime("%Y-%m-%d_%H-%M-%S")
+    args.seed = random.randint(1, 9999)
     os.makedirs("logs", exist_ok=True)
     logging.basicConfig(filename="logs/{}_{}_{}.log".format(os.path.basename(args.network),
-                                                            timestamp, args.seed), level=logging.INFO)
+                                                            DATETIME, args.seed), level=logging.INFO)
     logging.getLogger("coordsim").setLevel(logging.WARNING)
     logging.getLogger("bjointsp").setLevel(logging.INFO)
+
+    # Creating the results directory variable where the simulator result files will be written
+    network_stem = os.path.splitext(os.path.basename(args.network))[0]
+    service_function_stem = os.path.splitext(os.path.basename(args.service_functions))[0]
+    simulator_config_stem = os.path.splitext(os.path.basename(args.config))[0]
+
+    results_dir = f"{get_project_root()}/results/{network_stem}/{service_function_stem}/{simulator_config_stem}" \
+                  f"/{DATETIME}_seed{args.seed}"
     # creating the simulator
     # initializing the simulator with absolute paths to the network, service_functions and the config. files.
     simulator = Simulator(os.path.abspath(args.network),
                           os.path.abspath(args.service_functions),
-                          os.path.abspath(args.config), test_mode=True)
-
+                          os.path.abspath(args.config), test_mode=True, test_dir=results_dir)
     init_state = simulator.init(args.seed)
     log.info("Network Stats after init(): %s", init_state.network_stats)
     # assuming for now that there is only one SFC.
@@ -98,7 +106,7 @@ def main():
     # We run the simulator iterations number of times to get the Traffic info from the SimulatorAction object
     # We generate new source file for the BJointSP from the traffic info we get from the simulator for each iteration
     # Using this source file and the already generated Template file we call the 'place' fx of BJointSP
-    # Incase no source exists, we use the previous placement and schedule
+    # In-case no source exists, we use the previous placement and schedule
     for i in range(args.iterations):
         action = SimulatorAction(placement, schedule)
         apply_state = simulator.apply(action)
@@ -108,6 +116,9 @@ def main():
             result = bjointsp_place(os.path.abspath(args.network), os.path.abspath(template), os.path.abspath(source),
                                     cpu=node_cap, mem=node_cap, dr=1000, networkx=simulator.network, write_result=False)
             placement, schedule = get_placement_and_schedule(result, nodes_list, sfc_name, sf_list)
+
+    copy_input_files(results_dir, os.path.abspath(args.network), os.path.abspath(args.service_functions),
+                     os.path.abspath(args.config))
 
 
 if __name__ == '__main__':
